+++ src/MemeEditor/Terminal.cpp
@@ -117,7 +117,7 @@ namespace ml
 		ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(4, 1)); // Tighten spacing
 		
 		ImVec4 col_default_text = ImGui::GetStyleColorVec4(ImGuiCol_Text);
-		for (int32_t i = 0; i < m_lines.size(); i++)
+		for (size_t i = 0; i < m_lines.size(); i++)
 		{
 			CString item = m_lines[i].c_str();
 			if (!filter.PassFilter(item))
